pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-trader"
version = "0.3.3"
description = "Pandas-powered backtesting framework for algorithmic trading, featuring 20+ strategies, multi-market support, CLI tools, and an integrated MCP server for professional traders."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "GPL-3.0"}
authors = [
    {name = "Will Chien", email="locriginal@gmail.com"}
]
keywords = ["backtesting", "algorithmic-trading", "finance", "pandas", "technical-analysis", "mcp", "cli", "quantitative-finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=2.0.0",
    "scipy>=1.13.0",
    "pandas>=2.3.0",
    "yfinance>=0.2.40",
    "pandas-ta>=0.4.71b0",
    "peewee>=3.17.5",
    "matplotlib>=3.8.0",
    "contourpy>=1.2.0",
    "cycler>=0.12.1",
    "kiwisolver>=1.4.5",
    "fonttools>=4.60.0",
    "pillow>=10.3.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "html5lib>=1.1",
    "soupsieve>=2.5",
    "statsmodels>=0.14.2",
    "patsy>=0.5.6",
    "pydantic>=2.7",
    "pyyaml>=6.0.2",
    "python-dateutil>=2.9.0",
    "typing_extensions>=4.11.0",
    "tzdata>=2024.1",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-benchmark>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.4.8",
    "pre-commit>=3.0.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.31.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]
notebook = [
    "jupyter>=1.0.0",
    "nbval>=0.10.0",
    "ipykernel>=6.25.0",
]

[project.urls]
Homepage = "https://github.com/whchien/ai-trader"
Documentation = "https://ai-trader.readthedocs.io"
Repository = "https://github.com/whchien/ai-trader"
Issues = "https://github.com/whchien/ai-trader/issues"

[project.scripts]
ai-trader = "ai_trader.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_trader*"]
exclude = ["tests*", "docs*", "notebooks*"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable: pycodestyle errors (E), pyflakes (F), isort (I), pep8-naming (N),
# pycodestyle warnings (W), pyupgrade (UP)
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["E501"]  # Allow long lines in tests

[tool.ruff.lint.isort]
known-first-party = ["ai_trader"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=ai_trader",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["ai_trader"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Mypy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, will enable later
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "backtrader.*",
    "twstock.*",
    "yfinance.*",
    "statsmodels.*",
    "matplotlib.*",
]
ignore_missing_imports = true

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = ["B101"]  # Skip assert warnings (OK in tests)
